@@ -3,7 +3,7 @@ package com.example.payments.dao;
 import java.math.BigDecimal;
 import java.util.List;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 public class TaxesEntity {
